requirements.txt
aioredis==1.3.1
amqp==5.0.9
anyjson==0.3.3
asgiref==3.4.1
async-timeout==4.0.2
attrs==21.4.0
autobahn==21.11.1
Automat==20.2.0
billiard==3.6.4.0
celery==5.2.3
cffi==1.15.0
channels==3.0.4
channels-redis==3.3.1
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
constantly==15.1.0
cryptography==36.0.1
daphne==3.0.2
Deprecated==1.2.13
dj-database-url==0.5.0
Django==3.2.12
django-celery-beat==2.2.1
django-ckeditor==6.2.0
django-cors-headers==3.10.1
django-debug-toolbar==3.2.4
django-js-asset==1.2.2
django-timezone-field==4.2.3
django-webpack-loader==1.4.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
dnspython==2.2.0
greenlet==1.1.2
gunicorn==20.1.0
h11==0.13.0
hiredis==2.0.0
httptools==0.3.0
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
kombu==5.2.3
msgpack==1.0.3
packaging==21.3
Pillow==9.0.0
prompt-toolkit==3.0.26
psycopg2==2.9.3
psycopg2-binary==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.3.0
pyOpenSSL==21.0.0
pyparsing==3.0.7
python-crontab==2.6.0
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
PyYAML==6.0
redis==4.1.2
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.2
Twisted==21.7.0
txaio==21.2.1
typing-extensions==4.0.1
uvicorn==0.17.5
uvloop==0.16.0
vine==5.0.0
watchgod==0.7
wcwidth==0.2.5
websockets==10.1
wrapt==1.13.3
zope.interface==5.4.0
